requirements.txt
amqp==5.0.9
anyjson==0.3.3
asgiref==3.5.0
astroid==2.9.3
autopep8==1.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.0.0
click==8.0.3
click-didyoumean==0.3.0
click-repl==0.2.0
cloudpickle==2.0.0
dask==2022.1.1
Django==4.0.2
django-celery-beat==2.1.0
django-celery-results==2.0.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
fsspec==2022.1.0
isort==5.10.1
kombu==5.2.3
lazy-object-proxy==1.7.1
locket==0.2.1
mccabe==0.6.1
numpy==1.22.2
packaging==21.3
pandas==1.4.0
partd==1.2.0
platformdirs==2.4.1
prompt-toolkit==3.0.27
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
toolz==0.11.2
typing-extensions==4.0.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
